refactor : WinningLotto 생성자 파라미터 변경
