+++ src/main/java/lotto/controller/LottoAutoController.java
@@ -4,6 +4,7 @@ import lotto.model.*;
 import lotto.view.LottoAutoInputView;
 import lotto.view.LottoAutoResultView;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class LottoAutoController {
@@ -19,11 +20,28 @@ public class LottoAutoController {
         List<Lotto> lottos = autoLotto.setLotto(quantity);
         resultView.printLotto(quantity, lottos);
 
-        WinningLotto winningLotto = new WinningLotto(inputView.inputWinningNumbers(), inputView.inputBonusNumber());
+        List<Integer> winningNumbers = convertWinningNumbersToInt(inputView.inputWinningNumbers());
+        int bonusNumber = convertBonusNumberToInt(inputView.inputBonusNumber());
+        WinningLotto winningLotto = new WinningLotto(winningNumbers, bonusNumber);
 
         WinningResult winningResult = new WinningResult(lottos, winningLotto);
 
         resultView.outputWinningStatistic(getEarningRate(getEarningPrice(), LottoPrice.getPrice()));
+    }
+
+    public List<Integer> convertWinningNumbersToInt(String winningStringNumbers) {
+        List<Integer> winningNumbers = new ArrayList<>();
+        String[] splitWinningNumbers = winningStringNumbers.split(", ");
+        for (int i = 0; i < splitWinningNumbers.length; i++) {
+            int number = Integer.parseInt(splitWinningNumbers[i]);
+            winningNumbers.add(number);
+        }
+        return winningNumbers;
+    }
+
+    private int convertBonusNumberToInt(String bonusStringNumber) {
+        int bonusNumber = Integer.parseInt(bonusStringNumber);
+        return bonusNumber;
     }
 
     private float getEarningRate(int earningPrice, int purchasePrice) {
